@@ -3,6 +3,7 @@
   "version": "0.0.0",
   "private": "true",
   "scripts": {
+    "lint": "npx eslint packages/*/src/**/*.ts --fix",
     "clean-packages": "npx lerna run clean-packages && rm -rf node_modules",
     "build": "npx lerna run build --stream --include-dependencies",
     "build-and-test": "npm run build && npm run test",
@@ -39,16 +40,15 @@
     "@types/node": "^20.5.0",
     "@types/sinon": "^17.0.2",
     "@types/uuid": "^9.0.2",
-    "@typescript-eslint/eslint-plugin": "^5.62.0",
-    "@typescript-eslint/parser": "^5.55.0",
+    "@typescript-eslint/eslint-plugin": "^6.15.0",
     "@vitest/coverage-istanbul": "^1.0.4",
     "concurrently": "^8.2.0",
-    "eslint": "^8.47.0",
-    "eslint-config-next": "^12.0.0",
-    "eslint-config-standard-with-typescript": "^34.0.1",
-    "eslint-plugin-import": "^2.27.5",
-    "eslint-plugin-n": "^15.6.1",
+    "eslint": "^8.56.0",
+    "eslint-config-standard-with-typescript": "^43.0.0",
+    "eslint-plugin-import": "^2.29.1",
+    "eslint-plugin-n": "^16.5.0",
     "eslint-plugin-promise": "^6.1.1",
+    "eslint-plugin-unused-imports": "^3.0.0",
     "fake-indexeddb": "^4.0.2",
     "file-loader": "^6.2.0",
     "husky": "^7.0.4",
@@ -82,7 +82,7 @@
     "typedoc-plugin-rename-defaults": "^0.6.5",
     "typedoc-plugin-superstruct": "^1.0.0",
     "typedoc-theme-hierarchy": "^4.0.0",
-    "typescript": "^5.1.6",
+    "typescript": "^5.3.3",
     "vitest": "^1.0.4"
   },
   "dependencies": {
