@@ -1,56 +1,56 @@
-import { RxCollection, RxDocument, RxDocumentBase } from "rxdb";
-import type { GenericORMType, Schema } from "../types";
-import { Domain } from "@atala/prism-wallet-sdk";
+import { type RxCollection, type RxDocument } from 'rxdb'
+import type { Schema } from '../types'
+import { Domain } from '@atala/prism-wallet-sdk'
 
-export type MediatorSchemaType = {
-  id: string;
-  mediatorDID: string;
-  hostDID: string;
-  routingDID: string;
-};
+export interface MediatorSchemaType {
+  id: string
+  mediatorDID: string
+  hostDID: string
+  routingDID: string
+}
 
 const MediatorSchema: Schema<MediatorSchemaType> = {
   version: 0,
-  primaryKey: "id",
-  type: "object",
+  primaryKey: 'id',
+  type: 'object',
   properties: {
     id: {
-      type: "string",
-      maxLength: 60,
+      type: 'string',
+      maxLength: 60
     },
     mediatorDID: {
-      type: "string",
+      type: 'string'
     },
     hostDID: {
-      type: "string",
+      type: 'string'
     },
     routingDID: {
-      type: "string",
-    },
+      type: 'string'
+    }
   },
   encrypted: [],
-  required: ["id", "mediatorDID"],
-};
+  required: ['id', 'mediatorDID']
+}
 
-export type MediatorDocument = RxDocument<MediatorSchemaType>;
-export type MediatorMethodTypes = {
-  toDomainMediator: (this: RxDocument<MediatorSchemaType, MediatorMethodTypes>) => Domain.Mediator;
-};
+export type MediatorDocument = RxDocument<MediatorSchemaType>
+export interface MediatorMethodTypes {
+  toDomainMediator: (this: RxDocument<MediatorSchemaType, MediatorMethodTypes>) => Domain.Mediator
+}
 export type MediatorCollection = RxCollection<
-  MediatorSchemaType,
-  MediatorMethodTypes,
-  MediatorDocument
->;
+MediatorSchemaType,
+MediatorMethodTypes,
+MediatorDocument
+>
 
 export const MediatorMethods: MediatorMethodTypes = {
-  toDomainMediator: function toDomainMediator(this: RxDocument<MediatorSchemaType, MediatorMethodTypes>) {
+  toDomainMediator: function toDomainMediator (this: RxDocument<MediatorSchemaType, MediatorMethodTypes>) {
     const mediator = this.toJSON()
     return {
       hostDID: Domain.DID.fromString(mediator.hostDID),
       routingDID: Domain.DID.fromString(mediator.routingDID),
-      mediatorDID: Domain.DID.fromString(mediator.mediatorDID),
-    };
-  },
-};
+      mediatorDID: Domain.DID.fromString(mediator.mediatorDID)
+    }
+  }
+}
 
-export default MediatorSchema;
+export default MediatorSchema
