@@ -1,4 +1,4 @@
-export * from './rx-storage-dexie';
-export * from './rx-storage-instance-dexie';
-export * from './dexie-helper';
-export * from './dexie-query';
+export * from './rx-storage-dexie'
+export * from './rx-storage-instance-dexie'
+export * from './dexie-helper'
+export * from './dexie-query'
