@@ -4,7 +4,7 @@
  * @description This is a RXDB LevelDB storage that supports encryption middleware.
  * In order to use this in your pluto-encrypted database you must write the following code:
  * Creating a LevelDB compatible storage is very simple.
- * 
+ *
  * ```typescript
  * import { createLevelDBStorage } from "@pluto-encrypted/leveldb";
  * import { Database } from "@pluto-encrypted/database";
@@ -13,94 +13,87 @@
  * const database = db = await Database.createEncrypted({
  *          name: `my-db`,
  *          encryptionKey: defaultPassword,
- *          storage: createLevelDBStorage({ 
+ *          storage: createLevelDBStorage({
  *          dbName: "demo",
- *          dbPath: "/tmp/demo" 
+ *          dbPath: "/tmp/demo"
  *       })
  * });
  * ```
  */
-import { RxStorage, RxStorageDefaultStatics, RxStorageInstance, RxStorageInstanceCreationParams, newRxError } from "rxdb";
-import { LevelDBInternalConstructor, LevelDBSettings, LevelDBStorageInternals, RxStorageLevelDBType } from "./leveldb/types";
+import { RxStorageDefaultStatics, type RxStorageInstance, type RxStorageInstanceCreationParams, newRxError } from 'rxdb'
+import { type LevelDBInternalConstructor, type LevelDBSettings, type LevelDBStorageInternals, type RxStorageLevelDBType } from './leveldb/types'
 
-import { RxStorageIntanceLevelDB } from "./leveldb/instance";
-import { LevelDBInternal } from "./leveldb/internal";
-import { wrappedKeyEncryptionStorage } from "@pluto-encrypted/encryption";
+import { RxStorageIntanceLevelDB } from './leveldb/instance'
+import { LevelDBInternal } from './leveldb/internal'
+import { wrappedKeyEncryptionStorage } from '@pluto-encrypted/encryption'
 
-export type * from './leveldb/types';
+export type * from './leveldb/types'
 
-
-export type { MangoQuerySelectorAndIndex, CRDTSchemaOptions, RxQueryPlanKey, PrimaryKey, StringKeys, TopLevelProperty, MangoQuerySortDirection } from "rxdb/dist/types/types";
+export type { MangoQuerySelectorAndIndex, CRDTSchemaOptions, RxQueryPlanKey, PrimaryKey, StringKeys, TopLevelProperty, MangoQuerySortDirection } from 'rxdb/dist/types/types'
 export type { DefaultPreparedQuery, RxJsonSchema, FilledMangoQuery, CompressionMode, RxQueryPlan, MangoQuery, MangoQueryNoLimit, MangoQuerySortPart } from 'rxdb'
 export type { Level } from 'level'
-export const RX_STORAGE_NAME_LEVELDB = 'leveldb';
+export const RX_STORAGE_NAME_LEVELDB = 'leveldb'
 
-async function preloadData<RxDocType>(constructorProps: LevelDBInternalConstructor<RxDocType>) {
-    try {
-        const internalStorage = new LevelDBInternal<RxDocType>(constructorProps);
-        await internalStorage.getDocuments([]);
-        return internalStorage
-    } catch (err) {
-        throw err
-    }
+async function preloadData<RxDocType> (constructorProps: LevelDBInternalConstructor<RxDocType>) {
+  const internalStorage = new LevelDBInternal<RxDocType>(constructorProps)
+  await internalStorage.getDocuments([])
+  return internalStorage
 }
 
-let internalInstance: Map<string, LevelDBInternal<any>> = new Map()
+const internalInstance = new Map<string, LevelDBInternal<any>>()
 
-function getRxStorageLevel<RxDocType>(settings: LevelDBSettings): RxStorageLevelDBType<RxDocType> {
-    const instance: RxStorageLevelDBType<any> = {
-        name: RX_STORAGE_NAME_LEVELDB,
-        statics: RxStorageDefaultStatics,
-        async createStorageInstance<RxDocType>(params: RxStorageInstanceCreationParams<RxDocType, LevelDBSettings>): Promise<RxStorageInstance<RxDocType, LevelDBStorageInternals<RxDocType>, LevelDBSettings, any>> {
-            if (params.schema.keyCompression) {
-                throw newRxError('UT5', { args: { databaseName: params.databaseName, collectionName: params.collectionName } })
-            }
+function getRxStorageLevel<RxDocType> (settings: LevelDBSettings): RxStorageLevelDBType<RxDocType> {
+  const instance: RxStorageLevelDBType<any> = {
+    name: RX_STORAGE_NAME_LEVELDB,
+    statics: RxStorageDefaultStatics,
+    async createStorageInstance<RxDocType>(params: RxStorageInstanceCreationParams<RxDocType, LevelDBSettings>): Promise<RxStorageInstance<RxDocType, LevelDBStorageInternals<RxDocType>, LevelDBSettings, any>> {
+      if (params.schema.keyCompression) {
+        throw newRxError('UT5', { args: { databaseName: params.databaseName, collectionName: params.collectionName } })
+      }
 
-            const levelDBConstructorProps: LevelDBInternalConstructor<RxDocType> = "level" in settings ?
-                {
-                    level: settings.level,
-                    refCount: 1,
-                    schema: params.schema,
-                }
-                :
-                {
-                    dbPath: settings.dbPath,
-                    refCount: 1,
-                    schema: params.schema,
-                };
+      const levelDBConstructorProps: LevelDBInternalConstructor<RxDocType> = 'level' in settings
+        ? {
+            level: settings.level,
+            refCount: 1,
+            schema: params.schema
+          }
+        : {
+            dbPath: settings.dbPath,
+            refCount: 1,
+            schema: params.schema
+          }
 
-            const databasePath = "level" in levelDBConstructorProps ?
-                levelDBConstructorProps.level.location :
-                levelDBConstructorProps.dbPath;
+      const databasePath = 'level' in levelDBConstructorProps
+        ? levelDBConstructorProps.level.location
+        : levelDBConstructorProps.dbPath
 
-            const existingInstance = internalInstance.get(databasePath);
+      const existingInstance = internalInstance.get(databasePath)
 
-            if (!existingInstance) {
-                internalInstance.set(databasePath, await preloadData<RxDocType>(levelDBConstructorProps))
-            } else {
-                existingInstance.refCount++;
-                internalInstance.set(databasePath, existingInstance)
-            }
+      if (!existingInstance) {
+        internalInstance.set(databasePath, await preloadData<RxDocType>(levelDBConstructorProps))
+      } else {
+        existingInstance.refCount++
+        internalInstance.set(databasePath, existingInstance)
+      }
 
-            const rxStorageInstance = new RxStorageIntanceLevelDB<RxDocType>(
-                this,
-                params.databaseName,
-                params.collectionName,
-                params.schema,
-                internalInstance.get(databasePath)!,
-                settings
-            )
+      const rxStorageInstance = new RxStorageIntanceLevelDB<RxDocType>(
+        this,
+        params.databaseName,
+        params.collectionName,
+        params.schema,
+        internalInstance.get(databasePath)!,
+        settings
+      )
 
-            return rxStorageInstance
-        }
+      return rxStorageInstance
     }
+  }
 
-    return instance
+  return instance
 }
 
-
-export function createLevelDBStorage(settings: LevelDBSettings) {
-    return wrappedKeyEncryptionStorage({
-        storage: getRxStorageLevel(settings)
-    })
+export function createLevelDBStorage (settings: LevelDBSettings) {
+  return wrappedKeyEncryptionStorage({
+    storage: getRxStorageLevel(settings)
+  })
 }
